feat(web): show error popup in WorkSwitcher status badges
Extract shared WorkErrorPopup component and getSettingsLink helper so the
same error/warning popup (with settings deep-link) works in both the
WorkCard list and the WorkSwitcher sidebar dropdown.

WorkSwitcher: WorkStatusBadge now accepts an optional `work` prop and wraps
itself in HoverPopup when the work has a generation error, sync error, or
warnings — covering both the active-work badge in the trigger input and each
item row in the dropdown list.
